Fix transport trim settings persistence and improve speed slider UX
- Fixed transport trim (in/out points) not being respected during playback
- Added config persistence when trim values are updated via slider
- Added config persistence in _initialize_state() for both preserved and full-range trims
- Fixed get_chain() unconditionally resetting transport trim on every UI request
- Fixed apply_clip_effects() not checking for saved parameters before reinitializing
- Changed speed slider step from 0.01 to 0.1 for cleaner increments (0.1, 0.2, 0.3...)
- Updated speed value rounding to always snap to 0.1 increments for consistency

All trim settings now persist across page reloads, clip switches, and playlist changes.

+++ plugins/effects/radial_blur.py
@@ -98,13 +98,13 @@ class RadialBlurEffect(PluginBase):
                                     flags=cv2.INTER_LINEAR,
                                     borderMode=cv2.BORDER_REFLECT)
             
-            # Akkumuliere
-            result += zoomed.astype(np.float32)
+            # OPTIMIZED: Accumulate directly (auto-converts uint8 to float32)
+            result += zoomed
         
-        # Mittelwert bilden
-        result = result / iterations
+        # Average and convert back to uint8
+        result = (result / iterations).astype(np.uint8)
         
-        return result.astype(np.uint8)
+        return result
     
     def update_parameter(self, name, value):
         """Aktualisiert Parameter zur Laufzeit."""
